// src/infrastructure/grpc/generated/user/types/instructor_student.ts
export interface IsStudentOfInstructorResponse {
  success?: IsStudentOfInstructorSuccess | undefined;
  error?: Error | undefined;
}

export interface IsStudentOfInstructorSuccess {
  isStudent: boolean;
}
// ...
export const IsStudentOfInstructorResponse: MessageFns<IsStudentOfInstructorResponse> =
  {
    encode(
      message: IsStudentOfInstructorResponse,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      if (message.success !== undefined) {
        IsStudentOfInstructorSuccess.encode(
          message.success,
          writer.uint32(10).fork(),
        ).join();
      }
      if (message.error !== undefined) {
        Error.encode(message.error, writer.uint32(18).fork()).join();
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): IsStudentOfInstructorResponse {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseIsStudentOfInstructorResponse();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.success = IsStudentOfInstructorSuccess.decode(
              reader,
              reader.uint32(),
            );
            continue;
          }
          case 2: {
            if (tag !== 18) {
              break;
            }

            message.error = Error.decode(reader, reader.uint32());
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },
  };

function createBaseIsStudentOfInstructorSuccess(): IsStudentOfInstructorSuccess {
  return { isStudent: false };
}

export const IsStudentOfInstructorSuccess: MessageFns<IsStudentOfInstructorSuccess> =
  {
    encode(
      message: IsStudentOfInstructorSuccess,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      if (message.isStudent !== false) {
        writer.uint32(8).bool(message.isStudent);
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): IsStudentOfInstructorSuccess {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseIsStudentOfInstructorSuccess();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 8) {
              break;
            }

            message.isStudent = reader.bool();
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },
  };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
}